modelogestion/requirements.txt
djangorestframework
django-cors-headers
django-storages[boto3]
boto3
